BACKEND/requirements.txt
fastapi
nest_asyncio
pyngrok
uvicorn
pytesseract
Pillow
PyMuPDF
easyocr
sentence-transformers
faiss-cpu
numpy
groq
python-multipart